+++ ByBalance/Classes/Shared/BBBaseViewController.m
@@ -153,7 +153,7 @@
 {
     NSArray * stack = self.navigationController.viewControllers;
     
-    for (int i=stack.count-1; i > 0; --i)
+    for (NSInteger i=stack.count-1; i > 0; --i)
         if (stack[i] == self)
             return stack[i-1];
